templates.sifive: generate projects for all boards
- builds are functional
- hifive1 projects are functional
- arty projects to be tested

+++ bundles/ilg.gnumcueclipse.templates.sifive/templates/sifive_exe_c_project/system/riscv-arch/include/riscv-arch/csr-functions-inlines.h
@@ -344,7 +344,7 @@ extern "C"
       return tmp;
     }
 
-#endif
+#endif /* __riscv_xlen == 64 */
 
   static inline uint32_t
   __attribute__((always_inline))
@@ -365,9 +365,9 @@ extern "C"
 
 #elif __riscv_xlen == 64
 
-    return riscv_csr_read_mcycle ();
+    return (uint32_t)riscv_csr_read_mcycle ();
 
-#endif
+#endif /* __riscv_xlen */
   }
 
   static inline uint32_t
@@ -389,9 +389,9 @@ extern "C"
 
 #elif __riscv_xlen == 64
 
-    return (riscv_csr_read_mcycle() >> 32);
+    return (uint32_t)(riscv_csr_read_mcycle() >> 32);
 
-#endif
+#endif /* __riscv_xlen */
   }
 
 // ----------------------------------------------------------------------------
@@ -495,7 +495,7 @@ namespace riscv
         return riscv_csr_read_mcycle();
       }
 
-#endif
+#endif /* __riscv_xlen == 64 */
 
     inline uint32_t
     __attribute__((always_inline))
